backend/requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
flask-bcrypt
flask-jwt-extended
flask-session 